Layout link tests - A test for the links on the layout.
A test for the links on the layout. Now that we’ve filled in several of
the layout links, it’s a good idea to test them to make sure they’re
working correctly. We could do this by hand with a browser, first
visiting the root path and then checking the links by hand, but this
quickly becomes cumbersome. Instead, we’ll simulate the same series
of steps using an integration test, which allows us to write an
end-to-end test of our application’s behavior.