Tests for valid email formats. green
Note that we’ve included an optional second argument to the assertion
with a custom error message, which in this case identifies the address
causing the test to fail.

This uses the interpolated inspect method in section code.
Including the specific address that causes any failure is especially
useful in a test with an each loop like in section code, otherwise, any
failure would merely identify the line number, which is the same for all
the email addresses, and which wouldn’t be sufficient to identify the
source of the problem.