Validating the email format with a regular expression. green
Here the regex VALID_EMAIL_REGEX is a constant, indicated in Ruby by a
name starting with a capital letter.

In the commit ensures that only email addresses that match the pattern
will be considered valid. (The expression above has one minor weakness:
it allows invalid addresses that contain consecutive dots, such as
foo@bar..com. Updating the regex in Listing 6.21 to fix this blemish
is left as an exercise