A test for the rejection of duplicate email addresses. red
The method here is to make a user with the same email address as @user
using @user.dup, which creates a duplicate user with the same
attributes. Since we then save @user, the duplicate user has an email
address that already exists in the database, and hence should not be
valid.