The migration for enforcing email uniqueness.
This uses a Rails method called add_index to add an index on the email
column of the users table. The index by itself doesn’t enforce
uniqueness, but the option unique: true does.

The final step is to migrate the database.

If this fails, try exiting any running sandbox console sessions, which
can lock the database and prevent migrations.

In this case, I've removed the db/test.sqlite3 via:

$ rm -rf  db/test.sqlite3