An empty fixtures file. green
At this point, the test suite should be red due to a violation of the
uniqueness constraint in the fixtures, which contain sample data for
the test database. User fixtures were generated automatically
previous commits, and as the users.yml fixture email addresses are
not unique.

(They’re not valid either, but fixture data doesn’t get run
through the validations.)