Ensuring email uniqueness by downcasing the email attribute. green
The way to do this is with a callback, which is a method that gets
invoked at a particular point in the lifecycle of an Active Record
object. In the present case, that point is before the object is saved,
so we’ll use a before_save callback to downcase the email attribute
before saving the user.

(This is just a first implementation; we’ll discuss this subject again
in Section 11.1, where we’ll use the preferred method reference
convention for defining callbacks.)