The complete implementation for secure passwords. green
It’s good practice in general to enforce some minimum standards on
passwords to make them harder to guess. There are many options for
enforcing password strength in Rails, but for simplicity we’ll just
enforce a minimum length and the requirement that the password not be
blank. Picking a length of 6 as a reasonable minimum leads to the
validation test.