Adding some debug information to the site layout.
As preparation for adding dynamic pages to our sample application, now
is a good time to add some debug information to our site layout.
This displays some useful information about each page
using the built-in debug method and params variable.

To restrict the debug information to the development environment,
which is one of three environments defined by default in Rails.
In particular, Rails.env.development? is true only in a development
environment, so the embedded Ruby