The options hash used in the previous exercise is still commonly used, but as of Ruby 2.0 we can use keyword arguments instead.
Confirm that the code in this commit can be used in place of previous
commit.

What are the diffs between the two?

keyword argument is passed directly to the gravatar url with options
hash needs to be de-referenced and place into a local variable then
later passed to the gravatar url.