Exercises - replacing all occurrences of f with foobar
Confirm by replacing all occurrences of f with foobar that the name of
the block variable is irrelevant as far as the result is concerned.

Why might foobar nevertheless be a bad choice?

Foobar is not a good name variable name because it is used often in
application. It standard practice to label the variable block f or form.