A create action that can handle signup failure.
Our first step toward a working signup form is adding the code in
with this commit.

With this commit, it includes a second use of the render method, which
we first saw in the context of partials as you can see, render works in
controller actions as well. Note that we’ve taken this opportunity to
introduce an if-else branching structure, which allows us to handle the
cases of failure and success separately based on the value of @user.save
, which is either true or false depending on whether or not the save
succeeds.