Using strong parameters in the create action.
The comment included in  previous commit and reproduced above
indicates that this is not the final implementation. The reason is that
initializing the entire params hash is extremely dangerous—it arranges
to pass to User.new all data submitted by a user. In particular, suppose
that, in addition to the current attributes, the User model included an
admin attribute used to identify administrative users of the site.

To facilitate the use of these parameters, it’s conventional to
introduce an auxiliary method called user_params (which returns an
appropriate initialization hash) and use it in place of params[:user]:

Since user_params will only be used internally by the Users controller
and need not be exposed to external users via the web, we’ll make it
private using Ruby’s private keyword,