Code to display error messages on the signup form.
After running the console commands below

$ rails console
>> user = User.new(name: "Foo Bar", email: "foo@invalid",
?>                 password: "dude", password_confirmation: "dude")
>> user.save
=> false
>> user.errors.full_messages
=> ["Email is invalid", "Password is too short (minimum is 6 characters)
"]

We discover, that the errors.full_messages object contains an array of
error messages.

To display the messages in the browser, we’ll render an error-messages
partial on the user new page while adding the CSS class form-control
(which has special meaning to Bootstrap) to each entry field.