A partial for displaying form submission error messages.
This partial introduces several new Rails and Ruby constructs, including
two methods for Rails error objects. The first method is count, which
simply returns the number of errors.

The other new method is any?, which (together with empty?) is one of a
pair of complementary methods.

We see here that the empty? method, in the context of strings, also
works on Rails error objects, returning true for an empty object and
false otherwise. The any? method is just the opposite of empty?,
returning true if there are any elements present and false otherwise.

(By the way, all of these methods—count, empty?, and any?—work on Ruby
arrays as well. We’ll put this fact to good use starting in this commit.

The other new idea is the pluralize text helper, which is available in
the console via the helper object:

>> helper.pluralize(1, "error")
=> "1 error"
>> helper.pluralize(5, "error")
=> "5 errors"
We see here that pluralize takes an integer argument and then returns
the number with a properly pluralized version of its second argument.
Underlying this method is a powerful inflector that knows how to
pluralize a large number of words, including many with irregular
plurals:

>> helper.pluralize(2, "woman"),
=> "2 women"
>> helper.pluralize(3, "erratum")
=> "3 errata"

As a result of its use of pluralize, in this commit:

<%= pluralize(@user.errors.count, "error") %>