CSS for styling error messages.
The previous commit  includes the CSS id error_explanation for use in
styling the error messages. In addition, after an invalid submission
Rails automatically wraps the fields with errors in divs with the CSS
class field_with_errors.

These labels then allow us to style the error messages with the SCSS in
this commit which makes use of Sass’s @extend function to include the
functionality of the Bootstrap class has-error.

Now, helpful error messages now appear when submitting invalid signup
information.

Because the messages are generated by the model validations, they will
automatically change if you ever change your mind about, say, the format
of email addresses, or the minimum length of passwords.