Exercises - Change the minimum length of passwords to 5
Confirm by changing the minimum length of passwords to 5 that the error
message updates automatically as well

parameters: !ruby/hash:ActiveSupport::HashWithIndifferentAccess
     name: Carl
     email: carl_mcgruder@hotmail
     password: qqqq
     password_confirmation:

Results:

Password is too short (minimum is 5 characters)

How does the URL on the unsubmitted signup form compared
to the URL for a submitted signup form? Why don’t they match?

In the unsubmitted form the URL:/signup  and the submited form the
URL: /users.

The reason they do not match is because we GET the controller#action at
user#new that resolves to /sign but we POST to user#create that resolves
/users after the form is submitted.

Prefix    Verb    URI Patten                Controller#Action
-----------------------------------------------------------
          POST   /users(.:format)          users#create
signup    GET    /signup(.:format)         users#new