A test for an invalid signup. green
(Here User.count is 1 because of the user created in previous sections
though it may differ if you’ve added or deleted any users in the interim
.) In previous commits, we’ve used assert_select to test HTML elements
of the relevant pages, taking care to check only elements unlikely to
change in the future.

We’ll start by visiting the signup path using get:

get signup_path

In order to test the form submission, we need to issue a POST request to
the users_path, which we can do with the post function:

assert_no_difference 'User.count' do
post users_path, params: { user: {    name:  "",
                                     email: "user@invalid",
                                  password: "foo",
                     password_confirmation: "bar" } }
end

Here we’ve included the params[:user] hash expected by User.new in the
create action (Listing 7.29). (In versions of Rails before 5, params was
implicit, and only the user hash would be passed. This practice was
deprecated in Rails 5.0, and now the recommended method is to include
the full params hash explicitly.)

By wrapping the post in the assert_no_difference method with the string
argument ’User.count’, we arrange for a comparison between User.count
before and after the contents inside the assert_no_difference block.
This is equivalent to recording the user count, posting the data, and
verifying that the count is the same:

before_count = User.count
post users_path, ...
after_count  = User.count
assert_equal before_count, after_count
Although the two are equivalent, using assert_no_difference is cleaner
and is more idiomatically correct Ruby.

It’s worth noting that the get and post steps above are technically
unrelated, and it’s actually not necessary to get the signup path before
posting to the users path. I prefer to include both steps, though, both
for conceptual clarity and to double-check that the signup form renders
without error.

Putting the above ideas together leads to the test in in this commit.
We’ve also included a call to assert_template to check that a failed
submission re-renders the new action. Adding lines to check for the
appearance of error messages is left as an exercise.