Adding a signup route responding to POST requests.
The URLs for an unsubmitted signup form and for a submitted signup form
are /signup and /users, respectively, which don’t match. This is due to
our use of a custom named route in the former case previous commit
and a default RESTful route in the latter case next commit.
Resolve this discrepancy by adding the code shown in this commit.

Submit the new form to confirm that both cases now use the
same /signup URL. Are the tests still green? Why?

There are issues with current test because of css template added in
previous commit but after commenting out the inserted css template the
all the test passed.

The reason why we are still green is that we are using the
controller#action instead of using the named routes.