The user create action with a save and a redirect.
We can check the log and see that the create template error in the
server log.

No template found for UsersController#create, rendering head :no_content

Although it’s possible to render a template for the create action, the
usual practice is to redirect to a different page instead when the
creation is successful. In particular, we’ll follow the common
convention of redirecting to the newly created user’s profile, although
the root path would also work. The application code, which introduces
the redirect_to method, appears

Note that we’ve written

redirect_to @user

where we could have used the equivalent

redirect_to user_url(@user)