Iterating through a flash hash in the console.
(The key :success is a symbol, but embedded Ruby automatically converts
it to the string "success" before inserting it into the template.)

Using a different class for each key allows us to apply different styles
to different kinds of messages. For example, in future commit in Section
 8.1.4 we’ll use flash[:danger] to indicate a failed login attempt.

(In fact, we’ve already used alert-danger once, to style the error
message div in in this commit)

Bootstrap CSS supports styling for four such flash
classes for increasingly urgent message types (success, info, warning,
and danger), and we’ll find occasion to use all of them in the course
of developing the sample application (info in Section 11.2, warning
in Section 11.3, and danger for the first time in Section 8.1.4).