A test for a valid signup. green
In this case, we want to submit valid information and then confirm that
a user was created. In analogy with commit: A test for an invalid signup
, which used

assert_no_difference 'User.count' do
  post users_path, ...
end

here we’ll use the corresponding assert_difference method:

assert_difference 'User.count', 1 do
  post users_path, ...
end

As with assert_no_difference, the first argument is the string
’User.count’, which arranges for a comparison between User.count
before and after the contents of the assert_difference block. The second
 (optional) argument specifies the size of the difference (in this case,
1).

Incorporating assert_difference into the file from commit: A test for an
invalid signup, that yield sthe test shown in Listing 7.33. Note that
we’ve used the follow_redirect! method after posting to the users path.
This simplyarranges to follow the redirect after submission, resulting
in a rendering of the ’users/show’ template.