Generating the Sessions controller.
The elements of logging in and out correspond to particular REST actions
of the Sessions controller: the login form is handled by the new action
(covered in this section), actually logging in is handled by sending a
POST request to the create action, in up coming commits, and logging out
is handled by sending a DELETE request to the destroy action in up
coming commits.

Recall the association of HTTP verbs with REST actions from previous
commits

Including 'new' actually generates views as well, which is why we don’t
include actions like create and destroy that don’t correspond to views.