Adding a resource to get the standard RESTful actions for sessions. red
Unlike the Users resource, which used the special resources method to
obtain a full suite of RESTful routes automatically the Sessions
resource will use only named routes, handling GET and POST requests with
the login route and DELETE request with the logout route.

With the routes in this commit, we also need to update the test
generated in first commit with the new login route, as shown in next
commit.