Code for the login form.
Exercises - Submissions from the form

Submissions from the form defined in Listing 8.4 will be routed to the
Session controller’s create action.

How does Rails know to do this?

Because the form uses post, rails uses that via the routes to resolves
to login#create

Hint: Refer to table in the previous commit and the first line

Had to revert the config/puma.rb initial commit