An (unsuccessful) attempt at handling failed login.
These errors are associated with a particular Active Record object, but
this strategy won’t work here because the session isn’t an Active Record
model. Instead, we’ll put a message in the flash to be displayed upon
failed login.

Because of the flash message display in the site layout in previous
commit, the flash[:danger] message automatically gets displayed;
because of the Bootstrap CSS, it automatically gets nice styling.