The log_in function.
Because temporary cookies created using the session method are
automatically encrypted, the code in this commit is secure, and there
is no way for an attacker to use the session information to log in as
the user. This applies only to temporary sessions initiated with the
session method, though, and is not the case for persistent sessions
created using the cookies method. Permanent cookies are vulnerable to a
session hijacking attack, so in next branch we’ll have to be much more
careful about the information we place on the user’s browser.