Logging in a user.
With the log_in method defined in previous commit, we’re now ready to
complete the session create action by logging the user in and
redirecting to the user’s profile page.

With the create action defined in this commit, the login form defined
in this commit should now be working. It doesn’t have any effects on the
application display, though, so short of inspecting the browser session
directly there’s no way to tell that you’re logged in. As a first step
toward enabling more visible changes, in the commit with login form view
we’ll retrieve the current user from the database using the id in the
session.

In a future commit, we’ll change the links on the application layout,
including a URL to the current user’s profile.