The logged_in? helper method.
A user is logged in if there is a current user in the session, i.e., if
current_user is not nil. Checking for this requires the use of the “not”
operator (Section 4.2.3), written using an exclamation point ! and
usually read as “bang”. The resulting logged_in? method appears in
this commit.