Changing the layout links for logged-in users.
With the addition in this commit, we’re now ready to change the layout
links if a user is logged in. There are four new links, two of which are
stubbed out (to be completed in the branch: updating-users):

The logout link, meanwhile, uses the logout path previous commit:

<%= link_to "Log out", logout_path, method: :delete %>

Notice that the logout link passes a hash argument indicating that it
should submit with an HTTP DELETE request.7 We’ll also add a profile
link as follows:

<%= link_to "Profile", current_user %>

Here we could write

<%= link_to "Profile", user_path(current_user) %>

but as usual Rails allows us to link directly to the user by
automatically converting current_user into user_path(current_user)
in this context. Finally, when users aren’t logged in, we’ll use the
login path defined in Listing 8.2 to make a link to the login form:

<%= link_to "Log in", login_path %>