Adding the Bootstrap JavaScript library to application.js.
As part of including the new links into the layout, Listing 8.19 takes
advantage of Bootstrap’s ability to make dropdown menus. Note in
particular the inclusion of the special Bootstrap CSS classes such as
dropdown, dropdown-menu, etc. To activate the dropdown menu, we need to
include Bootstrap’s custom JavaScript library in the Rails asset
pipeline’s application.js file, as well as the jQuery library.

Do not to forget to updated the Gemfile to include: jquery-rails