Adding a digest method for use in fixtures.
There are several places we could put the resulting digest method, but
we’ll have an opportunity in future to reuse digest in the User
model. This suggests placing the method in user.rb. Because we won’t
necessarily have access to a user object when calculating the digest
(as will be the case in the fixtures file), we’ll attach the digest
method to the User class itself, which makes it a class method.