A test for user logging in with valid information. green
Although we’ve defined the password_digest attribute required by
has_secure_password, sometimes it’s convenient to refer to the plain
(virtual) password as well. Unfortunately, this is impossible to arrange
 with fixtures, and adding a password attribute to this commit causes
 Rails to complain that there is no such column in the database
 (which is true). We’ll make do by adopting the convention that all
 fixture users have the same password (’password’).

 Here users corresponds to the fixture filename users.yml, while the
 symbol :michael references user with the key

 Here we’ve used

assert_redirected_to @user

to check the right redirect target and

follow_redirect!

to actually visit the target page. Listing 8.23 also verifies that the
login link disappears by verifying that there are zero login path links
on the page:

assert_select "a[href=?]", login_path, count: 0

By including the extra count: 0 option, we tell assert_select that we
expect there to be zero links matching the given pattern.

(Compare this to count: 2 in following snippet, which checks for exactly
two matching links.)

Note that the assertion for the root path verifies that there are two
such links (one each for the logo and navigation menu element):

assert_select "a[href=?]", root_path, count: 2

This ensures that both links to the Home page defined are present.