Logging in the user upon signup.
Although our authentication system is now working, newly registered
users might be confused, as they are not logged in by default. Because
it would be strange to force users to log in immediately after signing
up, we’ll log in new users automatically as part of the signup process.

To arrange this behavior, all we need to do is add a call to log_in in
the Users controller create action