A boolean method for login status inside tests.
To test the behavior previous commit, we can add a line to the test
from from commit:A test for a valid signup
to check that the user is logged in. It’s helpful in
this context to define an is_logged_in? helper method to parallel the
logged_in? helper defined in commit: The logged_in? helper method.
, which returns true if there’s a user id in the (test) session and
false otherwise (commit: A boolean method for login status inside
tests.).

(Because helper methods aren’t available in tests, we can’t use the
current_user as in commit: The logged_in? helper method., but the
session method is available, so we use that instead.) Here we use
is_logged_in? instead oflogged_in? so that the test helper and Sessions
helper methods have different names, which prevents them from being
mistaken for each other.