The generated migration for the remember digest.
Information stored using session is automatically secure, but this is
not the case with information stored using cookies. In particular,
persistent cookies are vulnerable to session hijacking, in which an
attacker uses a stolen remember token to log in as a particular user.

There are four main ways to steal cookies:

  1. using a packet sniffer to detect cookies being passed over insecure
    networks,
  2. compromising a database containing remember tokens,
  3. using cross-site scripting (XSS)
  4. gaining physical access to a machine with a logged-in user.

We prevented the first problem in commit: Configuring the application to
use SSL in production, by using Secure Sockets Layer (SSL) site-wide,
which protects network data from packet sniffers. We’ll prevent the
second problem by storing a hash digest of the remember tokens instead
of the token itself, in much the same way that we stored password
digests instead of raw passwords in Section 6.3.2 - Adding a secure
password

Rails automatically prevents the third problem by escaping any content
inserted into view templates. Finally, although there’s no iron-clad
way to stop attackers who have physical access to a logged-in computer,
we’ll minimize the fourth problem by changing tokens every time a user
logs out and by taking care to cryptographically sign any potentially
sensitive information we place on the browser.

With these design and security considerations in mind, our plan for
creating persistent sessions appears as follows:

  1. Create a random string of digits for use as a remember token.
  2. Place the token in the browser cookies with an expiration date far
     in the future.
  3. Save the hash digest of the token to the database.
  4. Place an encrypted version of the user’s id in the browser cookies.

When presented with a cookie containing a persistent user id, find the
user in the database using the given id, and verify that the remember
token cookie matches the associated hash digest from the database.

Note how similar the final step is to logging a user in, where we
retrieve the user by email address and then verify (using the
authenticate method) that the submitted password matches the password
digest (commit: Finding and authenticating a user). As a result, our
implementation will parallel aspects of has_secure_password.

Do not forget to migrate application:

$ rails db:migrate