Adding a method for generating tokens.
Now we have to decide what to use as a remember token. There are many
mostly equivalent possibilities—essentially, any long random string will
do.

The urlsafe_base64 method from the SecureRandom module in the Ruby
standard library fits the bill:3 it returns a random string of length
22 composed of the characters A–Z, a–z, 0–9, “-”, and “_” (for a total
of 64 possibilities, thus “base64”)

Test this Ruby method in the console:

$ rails console
>> SecureRandom.urlsafe_base64
=> "q5lt38hQDc_959PVoo6b7A"

Just as it’s perfectly fine if two users have the same password, there’s
no need for remember tokens to be unique, but it’s more secure if they
are.

As a bonus, by using base64 strings specifically designed to be safe in
URLs (as indicated by the name urlsafe_base64), we’ll be able to use the
same token generator to make account activation and password reset
links in future commit.

Remembering users involves creating a remember token and saving the
digest of the token to the database. We’ve already defined a digest
method for use in the test fixtures (commit: Adding a digest method for
use in fixtures.), and we can use the results of the discussion above to
create a new_token method to create a new token. As with digest, the new
token method doesn’t need a user object, so we’ll make it a class method

The result is the User model in this commit.