Adding a remember method to the User model. green
Note the form of the assignment in the first line of the remember
method.
Because of the way Ruby handles assignments inside objects, without self
the assignment would create a local variable called remember_token,
which isn’t what we want.

Using self ensures that assignment sets the user’s remember_token
attribute. (Now you know why the before_save callback from commit:
Ensuring email uniqueness by downcasing the email attribute uses
self.email instead of just email.)

Meanwhile, the second line of remember uses the update_attribute method
to update the remember digest.

(As noted in in the following code snippet, this method bypasses the
validations, which is necessary in this case because we don’t have
access to the user’s password or confirmation.)

>> user.update_attributes(name: "The Dude", email: "dude@abides.org")
=> true
>> user.name
=> "The Dude"
>> user.email
=> "dude@abides.org

With these considerations in mind, we can create a valid token and
associated digest by first making a new remember token using
User.new_token, and then updating the remember digest with the result of
applying User.digest. This procedure gives the remember method shown in
this commit.

Exercises

In the console, assign user to the first user in the database, and
verify by calling it directly that the remember method works. How do
remember_token and remember_digest compare?

They do not compare, the remember_token is shorter while the
remember_digest is much longer plus the remember_token is a class
attribute while remember_digest is a stored db field is updated after
the remember method is called.