Defining the new token and digest methods using class << self. green
In commit: Adding a method for generating tokens, we defined the new
token and digest class methods by explicitly prefixing them with User.
This works fine and, because they are actually called using
User.new_token and User.digest, it is probably the clearest way to
define them. But there are two perhaps more idiomatically correct ways
to define class methods, one slightly confusing and one extremely
confusing. By running the test suite, verify that the implementations
in the commit: Defining the new token and digest methods using self
(slightly confusing) and in this commit (extremely confusing) are
correct. (Note that, in the context of in the commit: Defining the new
token and digest methods using self and in this commit, self is the
User class, whereas the other uses of self in the User model refer to a
user object instance. This is part of what makes them confusing.)