Adding an authenticated? method to the User model.
The final piece of the puzzle is to verify that a given remember token
matches the user’s remember digest, and in this context there are a
couple of equivalent ways to use bcrypt to verify a match. If you look
at the secure password source code, you’ll find a comparison like this:

BCrypt::Password.new(password_digest) == unencrypted_password

In our case, the analogous code would look like this:

BCrypt::Password.new(remember_digest) == remember_token

If you think about it, this code is really strange: it appears to be
comparing a bcrypt password digest directly with a token, which would
imply decrypting the digest in order to compare using ==. But the whole
point of using bcrypt is for hashing to be irreversible, so this can’t
be right. Indeed, digging into the source code of the bcrypt gem
verifies that the comparison operator == is being redefined, and under
the hood the comparison above is equivalent to the following:

BCrypt::Password.new(remember_digest).is_password?(remember_token)

Instead of ==, this uses the boolean method is_password? to perform the
comparison. Because its meaning is a little clearer, we’ll prefer this
second comparison form in the application code.

The above discussion suggests putting the digest–token comparison into
an authenticated? method in the User model, which plays a role similar
to that of the authenticate method provided by has_secure_password for
authenticating a user (commit:Logging in a user). The implementation
appears in this commit. (Although the authenticated? method in this
commit is tied specifically to the remember digest, it will turn out to
be useful in other contexts as well, and we’ll generalize it in Chapter
11.)

Note that the remember_token argument in the authenticated? method
defined in commit: Adding an authenticated? method to the User model.
is not the same as the accessor that we defined in commit: Adding a
remember method to the User model. using attr_accessor :remember_token;
instead, it is a variable local to the method. (Because the argument
refers to the remember token, it is not uncommon to use a method
argument that has the same name.) Also note the use of the
remember_digest attribute, which is the same as self.remember_digest
and, like name and email in Chapter 6, is created automatically by
Active Record based on the name of the corresponding database column
(commit: The generated migration for the remember digest).