Remembering the user.
As with log_in, commit: Logging in and remembering a user, defers the
real work to the Sessions helper, where we define a remember method
that calls user.remember, thereby generating a remember token and saving
its digest to the database. It then uses cookies to create permanent
cookies for the user id and remember token as described above.