Updating current_user for persistent sessions. red
With the code in commit: Remembering the user, a user logging in will
be remembered in the sense that their browser will get a valid remember
token, but it doesn’t yet do us any good because the current_user method
defined in commit: Finding the current user in the session
knows only about the temporary session:

@current_user ||= User.find_by(id: session[:user_id])

In the case of persistent sessions, we want to retrieve the user from
the temporary session if session[:user_id] exists, but otherwise we
should look for cookies[:user_id] to retrieve (and log in) the user
corresponding to the persistent session. We can accomplish this as
follows:

if session[:user_id]
  @current_user ||= User.find_by(id: session[:user_id])
elsif cookies.signed[:user_id]
  user = User.find_by(id: cookies.signed[:user_id])
  if user && user.authenticated?(cookies[:remember_token])
    log_in user
    @current_user = user
  end
end

(This follows the same user && user.authenticated pattern we saw in
commit: Finding and authenticating a user.)

The code above will work, but note the repeated use of both session and
cookies. We can eliminate this duplication as follows:

if (user_id = session[:user_id])
  @current_user ||= User.find_by(id: user_id)
elsif (user_id = cookies.signed[:user_id])
  user = User.find_by(id: user_id)
  if user && user.authenticated?(cookies[:remember_token])
    log_in user
    @current_user = user
  end
end

This uses the common but potentially confusing construction

if (user_id = session[:user_id])
Despite appearances, this is not a comparison (which would use double-
equals ==), but rather is an assignment. If you were to read it in words
, you wouldn’t say “If user id equals session of user id…”, but rather
something like “If session of user id exists (while setting user id to
session of user id)…”.

Defining the current_user helper as discussed above leads to the
implementation shown in this commit.

With the code as in this commit, newly logged in users are correctly
remembered, as you can verify by logging in, closing the browser, and
checking that you’re still logged in when you restart the sample
application and revisit the sample application. If you want, you can
even inspect the browser cookies to see the result directly.

There’s only one problem with our application as it stands: short of
clearing their browser cookies (or waiting 20 years), there’s no way
for users to log out. This is exactly the sort of thing our test suite
should catch, and indeed it should currently be red:

$ rails test

Reason why it red:

Failure:
UsersLoginTest#test_login_with_valid_information_followed_by_logout
[C:/Users/camcgruder/git/sample_app_5/test/integration/users_login_test
.rb:34]:
Expected at least 1 element matching "a[href="/login"]", found 0..
Expected 0 to be >= 1.

Exercises -

1. By finding the cookie in your local browser, verify that a remember
token and encrypted user id are present after logging in.

Yes, they are present

2. At the console, verify directly that the authenticated? method
defined in commit: Adding an authenticated? method to the User model,
works correctly.

Yes, it works correctly:

>> user = User.first
>> user.remembered
>> user.authenticated? user.remember_token
=> true