Logging out from a persistent session. green
With the code in commit: Adding a forget method to the User model,
we’re now ready to forget a permanent session by adding a forget helper
and calling it from the log_out helper in this commit. As seen iin this
commit, the forget helper calls user.forget and then deletes the
user_id and remember_token cookies.

At this point, the tests suite should be green:

$ rails test

After testing, I've discovered that VALID_EMAIL_REGEX was not
constructed properly so it failed validates process.

Exercises -

1. After logging out, verify that the corresponding cookies have been
removed from your browser.

Yes, I see that user_id and remember_token cookie has been created on
login and both as been deleted on logout.