A test for user logout. red
There are two closely related subtleties left to address. The first
subtlety is that, even though the “Log out” link appears only when
logged-in, a user could potentially have multiple browser windows open
to the site. If the user logged out in one window, thereby setting
current_user to nil, clicking the “Log out” link in a second window
would result in an error because of forget(current_user) in the
log_out method (commit: Logging out from a persistent session).
We can avoid this by logging out only if the user is logged in.

The second subtlety is that a user could be logged in (and remembered)
in multiple browsers, such as Chrome and Firefox, which causes a
problem if the user logs out in the first browser but not the second,
and then closes and re-opens the second one. For example, suppose
that the user logs out in Firefox, thereby setting the remember digest
to nil (via user.forget in commit: Adding a forget method to the User
model). The application will still work in Firefox; because the log_out
method in the commit: Logging out from a persistent session deletes the
user’s id, both highlighted conditionals are false.

As a result, evaluation falls off the end of the current_user method,
thereby returning nil as required.

In contrast, if we close Chrome, we set session[:user_id] to nil
(because all session variables expire automatically on browser close),
but the user_id cookie will still be present. This means that the
corresponding user will still be pulled out of the database when Chrome
is re-launched:

def current_user
  if (user_id = session[:user_id])
    @current_user ||= User.find_by(id: user_id)
  elsif (user_id = cookies.signed[:user_id])
    user = User.find_by(id: user_id)
    if user && user.authenticated?(cookies[:remember_token])
      log_in user
      @current_user = user
    end
  end
end

Consequently, the inner if conditional will be evaluated:

user && user.authenticated?(cookies[:remember_token])
In particular, because user isn’t nil, the second expression will be
evaluated, which raises an error. This is because the user’s remember
digest was deleted as part of logging out (commit: Adding a forget
method to the User model) in Firefox, so when we access the application
in Chrome we end up calling

BCrypt::Password.new(remember_digest).is_password?(remember_token)

with a nil remember digest, thereby raising an exception inside the
bcrypt library. To fix this, we want authenticated? to return false
instead.

These are exactly the sorts of subtleties that benefit from test-driven
development, so we’ll write tests to catch the two errors before
correcting them. We first get the integration test from commit: A test
for user logout. to red, as shown in this commit.

The second call to delete logout_path in Listing 9.14 should raise an
error due to the missing current_user, leading to a red test suite:

$ rails test

Error:
UsersLoginTest#test_login_with_valid_information_followed_by_logout:
NoMethodError: undefined method `forget' for nil:NilClass
    app/helpers/sessions_helper.rb:34:in `forget'
    app/helpers/sessions_helper.rb:41:in `log_out'
    app/controllers/sessions_controller.rb:18:in `destroy'
    test/integration/users_login_test.rb:34:in `block in
                                                <class:UsersLoginTest>'